working on parsing data for  weapons from api
