+++ spec/requests/github_services_spec.rb
@@ -7,39 +7,41 @@ describe GithubService do
         search = GithubService.get_units_by_faction("Necrons.cat")
 
         expect(search).to be_a Hash
-        objects = search[:catalogue][:sharedSelectionEntries][:selectionEntry].flat_map do |object|
-          if object[:type] == "model"
-            # binding.pry
+        expect(search[:catalogue][:sharedSelectionEntries][:selectionEntry]).to be_an Array
 
-            name = object[:name]
+        # objects = search[:catalogue][:sharedSelectionEntries][:selectionEntry].flat_map do |object|
+        #   if object[:type] == "model"
+        #     # binding.pry
 
-            stats = object[:profiles][:profile].flat_map do |entry|
-              if entry[:typeName] == "Unit"
-                entry[:characteristics][:characteristic]
-              end
-            end.compact
+        #     name = object[:name]
 
-            # weapons = object[:selectionEntries][:selectionEntry]
-            binding.pry
-            melee_weapons = object[:selectionEntryGroups][:selectionEntryGroup][:selectionEntries][:selectionEntry][:selectionEntryGroups][:selectionEntryGroup][:selectionEntries][:selectionEntry].select do |entry|
-              entry[:profiles] && entry[:profiles][:profile][:typeName] == "Melee Weapons"
-            end
+        #     stats = object[:profiles][:profile].flat_map do |entry|
+        #       if entry[:typeName] == "Unit"
+        #         entry[:characteristics][:characteristic]
+        #       end
+        #     end.compact
+
+        #     # weapons = object[:selectionEntries][:selectionEntry]
+        #     binding.pry
+        #     melee_weapons = object[:selectionEntryGroups][:selectionEntryGroup][:selectionEntries][:selectionEntry][:selectionEntryGroups][:selectionEntryGroup][:selectionEntries][:selectionEntry].select do |entry|
+        #       entry[:profiles] && entry[:profiles][:profile][:typeName] == "Melee Weapons"
+        #     end
             
-            # .map do |entry|
-            #   {
-            #     name: entry[:profiles][:profile][:name],
-            #     stats: entry[:profiles][:profile][:characteristics][:characteristic]
-            #   }
-            # end
-            {
-              :name => name,
-              :stats => stats,
-              :weapons => melee_weapons
-            }
-          end
-        end.compact
-        binding.pry
-        expect(search[:results]).to be_an Array
+        #     # .map do |entry|
+        #     #   {
+        #     #     name: entry[:profiles][:profile][:name],
+        #     #     stats: entry[:profiles][:profile][:characteristics][:characteristic]
+        #     #   }
+        #     # end
+        #     {
+        #       :name => name,
+        #       :stats => stats,
+        #       :weapons => melee_weapons
+        #     }
+        #   end
+        # end.compact
+        # binding.pry
+
 
       end
     end
